Assignment A
Create a webpage on which you can search for movie trailers
1. 	Use an API of an online movie database (e.g. IMDB or Rotten Tomatoes);
2. 	Use an API of an online video service (e.g. YouTube or Vimeo);
3. 	Create your own WebAPI as middleware to retrieve the results of both services and aggregate them;

I have created a rest webservice which fetches all upcoming movies and its corresponding trailer videos. The Rest Service sends back custom created json data with minimal information to the webpage. 
In the web page, once the page is loaded all the movies data fetched by the rest service is sent using Angular service. The data is loaded in tabular format and its trailer is embedded on the page.